+++ Editor/Settings/SettingsBase.cs
@@ -1,0 +1,15 @@
+namespace SOSXR.SeaShark
+{
+    public abstract class SettingsBase
+    {
+        public abstract void OnGUI();
+    }
+
+    public abstract class ProjectSettingsBase : SettingsBase
+    {
+    }
+
+    public abstract class PreferencesBase : SettingsBase
+    {
+    }
+}